pyproject.toml
[project]
name = "pymidi"
version = "0.1.0"
description = "Control your computer with a MIDI or HID Device."
requires-python = ">=3.13"
dependencies = [
    "hid>=1.0.8",
    "mido>=1.3.3",
    "pillow==12.0.0",
    "pyautogui>=0.9.54",
    "python-rtmidi==1.5.8",
]
